Add a chat message parser
A class to parse the chat message for special content. Internally keeps a list of objects that conform to a JSON content extractable protocol which can build the special content JSON incrementally.

This is so that if new content extractors are added, it can be added with ease. Also, if the existing content extractors need to be changed internally, this class does not need to know about those changes