Add a link extractor
Extracts links from a string. It uses a simple regex pattern to get the links but of course there are better ways to do this. If the implementation needs to change, it can be done without the users of this class knowing what those changes were.

Returns a list of dictionaries where each dictionary contains the link and the title of the link.